Remove about menu
We're removing all links to the views on this page but the postcards
view. In order to keep the postcards page alive, we're going to
remove the menu from it.

This is hacky for now but for the sake of minimizing the already
uncomfortably large scope of the about pages, I'm not going to
clean this up until after the they get deployed.