Begin gutting the about pages
A start on removing all the dead code from this repo, of
which there is a lot now. This isn't as complete as it
could be to limit the scope of the deletions. Plus I'm not
sure what this module should be called anymore so I stopped
short of variable renaming/handling a different namespace.